split scores into two files
